fix progress inconsistency after rebalance
Whenever there is a rebalance after increasing a topic's partition number to a higher value, there will be an interval of time where the consumer will not see the new partitions.

To prevent an error while monitoring the consumption progress, we now ignore the topic partitions until the consumer sees them.